Rename api.apis to api.api_classes
