+++ lineapy/api/api_classes.py
@@ -48,15 +48,14 @@ class LineaArtifact:
     name: str
     """name of the artifact"""
     _version: int
-    """version of the artifact - This is set when the artifact is saved. The format of the version currently is specified by the constant :const:`lineapy.utils.constants.VERSION_DATE_STRING`"""
+    """version of the artifact - currently start from 0"""
     date_created: Optional[datetime] = field(default=None, repr=False)
     # setting repr to false for date_created for now since it duplicates version
     """Optional because date_created cannot be set by the user. 
-    it is supposed to be automatically set when the 
-    artifact gets saved to the db. so when creating lineaArtifact 
-    the first time, it will be unset. When you get the artifact or 
-    list of artifacts as an artifact store, we retrieve the date 
-    from db and it will be set."""
+    it is supposed to be automatically set when the artifact gets saved to the 
+    db. so when creating lineaArtifact the first time, it will be unset. When 
+    you get the artifact or list of artifacts as an artifact store, we retrieve 
+    the date from db directly"""
 
     @property
     def version(self) -> int:
